refactor: independ read&gen font unit
